+++ src/main/java/com/yatensoft/dcbot/component/impl/fab/FabFabWebsiteParserImpl.java
@@ -1,9 +1,9 @@
 /** By YamiY Yaten */
-package com.yatensoft.dcbot.fab.component.impl;
+package com.yatensoft.dcbot.component.impl.fab;
 
 import com.yatensoft.dcbot.constant.WebsiteParserConstant;
-import com.yatensoft.dcbot.fab.component.skeleton.FabWebsiteParser;
-import com.yatensoft.dcbot.fab.dto.FabPagesListBlockDTO;
+import com.yatensoft.dcbot.component.skeleton.fab.FabWebsiteParser;
+import com.yatensoft.dcbot.dto.fab.FabPagesListBlockDTO;
 import java.io.IOException;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -20,6 +20,13 @@ import org.springframework.stereotype.Component;
  */
 @Component
 public class FabFabWebsiteParserImpl implements FabWebsiteParser {
+    /** fabtcg.com Constants */
+    private final String FABTCG_PAGES_LIST_BLOCK_ID = "pagesListblock";
+
+    private final String FABTCG_ITEM_LINK_CLASS = "item-link";
+    private final String FABTCG_URL_ELEMENT_CSS_QUERY = "a.item-link";
+    private final String FABTCG_TITLE_ELEMENT_CSS_QUERY = ".item-card div.item-card-inner h5";
+    private final String FABTCG_DATE_ELEMENT_CSS_QUERY = ".item-card div.item-card-inner p";
     @Value("${fabtcg.url.articles}")
     private String FABTCG_ARTICLES_URL;
 
@@ -30,9 +37,9 @@ public class FabFabWebsiteParserImpl implements FabWebsiteParser {
             /* Get document */
             final Document doc = Jsoup.connect(FABTCG_ARTICLES_URL).get();
             /* Get pagesListBlock element by its ID */
-            final Element pagesListBlock = doc.getElementById(WebsiteParserConstant.FABTCG_PAGES_LIST_BLOCK_ID);
+            final Element pagesListBlock = doc.getElementById(FABTCG_PAGES_LIST_BLOCK_ID);
             /* Get all item elements */
-            final Elements items = pagesListBlock.getElementsByClass(WebsiteParserConstant.FABTCG_ITEM_LINK_CLASS);
+            final Elements items = pagesListBlock.getElementsByClass(FABTCG_ITEM_LINK_CLASS);
             return items.stream()
                     .map(item -> item.attr(WebsiteParserConstant.HREF))
                     .findFirst()
@@ -42,17 +49,11 @@ public class FabFabWebsiteParserImpl implements FabWebsiteParser {
         }
     }
 
-    /** See {@link FabWebsiteParser#getLatestBanListUrl()} */
-    @Override
-    public String getLatestBanListUrl() throws IOException {
-        return null;
-    }
-
     private FabPagesListBlockDTO parseHTMLToItem(Element element) {
         FabPagesListBlockDTO item = new FabPagesListBlockDTO();
-        Element linkElement = element.selectFirst(WebsiteParserConstant.FABTCG_URL_ELEMENT_CSS_QUERY);
-        Element titleElement = element.selectFirst(WebsiteParserConstant.FABTCG_TITLE_ELEMENT_CSS_QUERY);
-        Element dateElement = element.selectFirst(WebsiteParserConstant.FABTCG_DATE_ELEMENT_CSS_QUERY);
+        Element linkElement = element.selectFirst(FABTCG_URL_ELEMENT_CSS_QUERY);
+        Element titleElement = element.selectFirst(FABTCG_TITLE_ELEMENT_CSS_QUERY);
+        Element dateElement = element.selectFirst(FABTCG_DATE_ELEMENT_CSS_QUERY);
 
         if (linkElement != null) {
             item.setLink(linkElement.attr(WebsiteParserConstant.HREF));
