+++ src/main/java/com/yatensoft/dcbot/component/impl/ygo/YgoScheduledTaskImpl.java
@@ -1,5 +1,5 @@
 /** By YamiY Yaten */
-package com.yatensoft.dcbot.ygo.component.impl;
+package com.yatensoft.dcbot.component.impl.ygo;
 
 import com.yatensoft.dcbot.config.DiscordBotConfig;
 import com.yatensoft.dcbot.constant.ChannelConstant;
@@ -7,10 +7,11 @@ import com.yatensoft.dcbot.constant.MessageConstant;
 import com.yatensoft.dcbot.dto.UrlArchiveDTO;
 import com.yatensoft.dcbot.enumeration.ArchiveTypeEnum;
 import com.yatensoft.dcbot.enumeration.TopicEnum;
-import com.yatensoft.dcbot.fab.component.skeleton.FabWebsiteParser;
 import com.yatensoft.dcbot.service.skeleton.UrlArchiveService;
-import com.yatensoft.dcbot.ygo.component.skeleton.YgoScheduledTask;
+import com.yatensoft.dcbot.component.skeleton.ygo.YgoScheduledTask;
 import java.io.IOException;
+
+import com.yatensoft.dcbot.component.skeleton.ygo.YgoWebsiteParser;
 import net.dv8tion.jda.api.entities.channel.concrete.TextChannel;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -22,7 +23,7 @@ import org.springframework.stereotype.Component;
  */
 @Component
 public class YgoScheduledTaskImpl implements YgoScheduledTask {
-    private final FabWebsiteParser websiteParser;
+    private final YgoWebsiteParser websiteParser;
     private final UrlArchiveService urlArchiveService;
 
     public YgoScheduledTaskImpl(
