+++ src/main/java/com/yatensoft/dcbot/component/skeleton/fab/FabScheduledTask.java
@@ -1,5 +1,5 @@
 /** By YamiY Yaten */
-package com.yatensoft.dcbot.fab.component.skeleton;
+package com.yatensoft.dcbot.component.skeleton.fab;
 
 import java.io.IOException;
