+++ src/main/java/com/yatensoft/dcbot/component/skeleton/fab/FabWebsiteParser.java
@@ -1,5 +1,5 @@
 /** By YamiY Yaten */
-package com.yatensoft.dcbot.fab.component.skeleton;
+package com.yatensoft.dcbot.component.skeleton.fab;
 
 import java.io.IOException;
 
@@ -13,11 +13,4 @@ public interface FabWebsiteParser {
      * @throws IOException if unable to parse website
      */
     String getLatestArticleUrl() throws IOException;
-
-    /**
-     * Get latest ban-list URL
-     * @return url of the last article
-     * @throws IOException if unable to parse website
-     */
-    String getLatestBanListUrl() throws IOException;
 }
