+++ src/main/java/com/yatensoft/dcbot/constant/WebsiteParserConstant.java
@@ -2,14 +2,6 @@
 package com.yatensoft.dcbot.constant;
 
 public class WebsiteParserConstant {
-    /** fabtcg.com Constants */
-    public static final String FABTCG_PAGES_LIST_BLOCK_ID = "pagesListblock";
-
-    public static final String FABTCG_ITEM_LINK_CLASS = "item-link";
-    public static final String FABTCG_URL_ELEMENT_CSS_QUERY = "a.item-link";
-    public static final String FABTCG_TITLE_ELEMENT_CSS_QUERY = ".item-card div.item-card-inner h5";
-    public static final String FABTCG_DATE_ELEMENT_CSS_QUERY = ".item-card div.item-card-inner p";
-
     /** yugioh-card.com Constants */
     public static final String YGOCARD_BLOCK_BUTTON_CLASS = "wp-block-button__link";
