+++ src/main/java/com/yatensoft/dcbot/orchestrator/impl/ygo/YgoCommandOrchestrator.java
@@ -1,10 +1,10 @@
 /** By YamiY Yaten */
-package com.yatensoft.dcbot.ygo.orchestrator;
+package com.yatensoft.dcbot.orchestrator.impl.ygo;
 
 import com.yatensoft.dcbot.constant.BotCommandConstant;
 import com.yatensoft.dcbot.constant.MessageConstant;
 import com.yatensoft.dcbot.orchestrator.skeleton.CommandOrchestrator;
-import com.yatensoft.dcbot.ygo.service.skeleton.YgoCommandService;
+import com.yatensoft.dcbot.service.skeleton.ygo.YgoCommandService;
 import java.io.IOException;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import org.springframework.beans.factory.annotation.Autowired;
