+++ src/main/java/com/yatensoft/dcbot/service/skeleton/ygo/YgoCommandService.java
@@ -1,5 +1,5 @@
 /** By YamiY Yaten */
-package com.yatensoft.dcbot.ygo.service.skeleton;
+package com.yatensoft.dcbot.service.skeleton.ygo;
 
 import java.io.IOException;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
